Refactoring of send_message model
